video_splicer/requirements.txt
google-cloud-storage
opencv-python-headless
numpy
